feat: add ts extention for user service
